redcap-data-dashboard/requirements.txt
dash
pandas
python-dotenv
requests
plotly
